Add 2.13-18
add 2.13

fix: dropdown links in 2.13

add 2.14

add 2.15

Fix 2.15 'c' div positioning

This is a discrepancy from the original code. In this page's code, the 'body' element needs to be given a relative position so the div tags can be absolutely positioned correctly. So, we need to be redundant in giving 'c' its absolute positioning here.

Add 2.16

Add 2.18